+++ packages/encryption/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.6](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.5...@standardnotes/encryption@1.4.6) (2022-04-20)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+
+
+
+
+## [1.4.5](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.4...@standardnotes/encryption@1.4.5) (2022-04-20)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+
+
+
+
+## [1.4.4](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.3...@standardnotes/encryption@1.4.4) (2022-04-19)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+
+
+
+
+## [1.4.3](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.2...@standardnotes/encryption@1.4.3) (2022-04-19)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+
+
+
+
+## [1.4.2](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.1...@standardnotes/encryption@1.4.2) (2022-04-19)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
+
+
+
+
 ## [1.4.1](https://github.com/standardnotes/snjs/compare/@standardnotes/encryption@1.4.0...@standardnotes/encryption@1.4.1) (2022-04-18)